requirements.txt
asgiref==3.9.2
certifi==2025.8.3
charset-normalizer==3.4.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.4.0
Django==5.2.6
django-rest-swagger==2.2.0
djangorestframework==3.16.1
filelock==3.19.1
idna==3.11
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
openapi-codec==1.3.2
packaging==25.0
peppercorn==0.6
pipenv==2025.0.4
platformdirs==4.4.0
requests==2.32.5
sampleproject==4.0.0
setuptools==80.9.0
simplejson==3.20.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.34.0
